+++ sources/Interop/Windows/other/helper-types/AUTHZ_CLIENT_CONTEXT_HANDLE.cs
@@ -51,11 +51,11 @@ namespace TerraFX.Interop
 
         public static explicit operator AUTHZ_CLIENT_CONTEXT_HANDLE(long value) => new AUTHZ_CLIENT_CONTEXT_HANDLE((nint)(value));
 
-        public static implicit operator long(AUTHZ_CLIENT_CONTEXT_HANDLE value) => (long)(value.Value);
+        public static implicit operator long(AUTHZ_CLIENT_CONTEXT_HANDLE value) => value.Value;
 
         public static explicit operator AUTHZ_CLIENT_CONTEXT_HANDLE(nint value) => new AUTHZ_CLIENT_CONTEXT_HANDLE((nint)(value));
 
-        public static implicit operator nint(AUTHZ_CLIENT_CONTEXT_HANDLE value) => (nint)(value.Value);
+        public static implicit operator nint(AUTHZ_CLIENT_CONTEXT_HANDLE value) => value.Value;
 
         public static explicit operator AUTHZ_CLIENT_CONTEXT_HANDLE(sbyte value) => new AUTHZ_CLIENT_CONTEXT_HANDLE((nint)(value));
